switched to p2 physics for circles etc

+++ ui/app/scripts/controllers/main.js
@@ -34,52 +34,100 @@ angular.module('uiApp')
             preload: function () {
                 this.load.tilemap('testmaze', 'assets/tilemaps/testoutdoor.json', null, Phaser.Tilemap.TILED_JSON);
 
-                this.load.spritesheet('hyptosis_tile-art-batch-1', 'images/hyptosis_tile-art-batch-1.png', 32,32);
+                this.load.spritesheet('hyptosis_tile-art-batch-1', 'images/hyptosis_tile-art-batch-1.png', 32, 32);
                 this.load.image('player', 'images/HB_Dwarf05.png');
                 this.load.image('demon', 'images/DemonMinorFighter.png');
             },
 
             create: function () {
+                this.game.physics.startSystem(Phaser.Physics.P2JS);
+
                 var map = this.game.add.tilemap('testmaze');
                 map.addTilesetImage('hyptosis_tile-art-batch-1');
 
                 this.blockLayer = map.createLayer('Block Layer');
                 //  TODO - darken
-                //this.blockLayer.tint = 0x00264d;
-                map.createLayer('Path');//.tint = 0x00264d;
+                this.blockLayer.tint = 0x1f1b34;
+                map.createLayer('Path').tint = 0x1f1b34;
                 this.blockLayer.resizeWorld();
                 map.setCollision([574, 575, 208, 79, 142, 146, 177]);
 
 
-                this.game.physics.startSystem(Phaser.Physics.ARCADE);
+                this.game.physics.p2.convertTilemap(map, this.blockLayer);
+                this.game.physics.p2.setBoundsToWorld(true, true, true, true, false);
+
+                //  TODO - physics for all objects really
+                //  https://code.google.com/p/box2d-editor/
+                //  http://phaser.io/examples/v2/p2-physics/load-polygon-1
 
                 this.player = this.game.add.sprite(16, 1264, 'player');
                 this.player.height = 32;
                 this.player.width = 32;
                 this.player.anchor.set(0.5);
-                this.game.physics.arcade.enable(this.player, Phaser.Physics.ARCADE, true);
-                this.player.body.setSize(24, 28, 0, 0);
+                this.game.physics.p2.enable(this.player);
+                this.player.body.setCircle(10);
                 this.player.body.collideWorldBounds = true;
+                this.player.body.mass = 10;
+                this.player.body.fixedRotation = true;
+                this.player.tint = 0x1f1b34;
+
+                var playerMaterial = game.physics.p2.createMaterial('playerMaterial', this.player.body);
+                var worldMaterial = game.physics.p2.createMaterial('worldMaterial');
+                var rockMaterial = game.physics.p2.createMaterial('rockMaterial');
+                var demonMaterial = game.physics.p2.createMaterial('demonMaterial');
+
+                game.physics.p2.setWorldMaterial(worldMaterial, true, true, true, true);
+                this.game.physics.p2.createContactMaterial(playerMaterial, worldMaterial, {
+                    friction: 0.01,
+                    restitution: 1,
+                    stiffness: 0
+                });
+                this.game.physics.p2.createContactMaterial(rockMaterial, worldMaterial, {
+                    friction: 0.9,
+                    restitution: 0.1,
+                    stiffness: 1e7,
+                    relaxation: 3,
+                    frictionStiffness: 1e7,
+                    frictionRelaxation: 3,
+                    surfaceVelocity: 0
+                });
+                this.game.physics.p2.createContactMaterial(demonMaterial, worldMaterial, {
+                    friction: 0,
+                    restitution: 1,
+                    stiffness: 0,
+                    relaxation: 0,
+                    frictionStiffness: 0,
+                    frictionRelaxation: 0,
+                    surfaceVelocity: 0
+                });
+                this.game.physics.p2.createContactMaterial(demonMaterial, rockMaterial, {
+                    friction: 1.0,
+                    restitution: 0.0,
+                    stiffness: 1e7,
+                    relaxation: 3,
+                    frictionStiffness: 1e7,
+                    frictionRelaxation: 3,
+                    surfaceVelocity: 0
+                });
                 this.game.camera.follow(this.player);
 
-                this.enemyGroup = this.game.add.group();
-                this.enemyGroup.enableBody = true;
+                this.game.physics.p2.createCollisionGroup();
+                this.enemyGroup = this.game.add.physicsGroup(Phaser.Physics.P2JS);
 
-                this.rockGroup = this.game.add.group();
-                this.rockGroup.enableBody = true;
-                //  TODO - would work better as a hexagon I think
+                this.rockGroup = this.game.add.physicsGroup(Phaser.Physics.P2JS);
+
                 map.createFromObjects('Object Layer', 214, 'hyptosis_tile-art-batch-1', 214, true, false, this.rockGroup);
                 angular.forEach(this.rockGroup.children, function (rock) {
                     rock.height = 40;
                     rock.width = 40;
-                    rock.anchor.set(0.5);
-                    this.game.physics.arcade.enable(rock, Phaser.Physics.ARCADE, true);
-                    rock.body.setSize(28, 32, 0, 0);
-                    rock.body.bounce.set(0.1);
+                    //rock.body.debug = true;
+                    rock.body.setRectangle(40, 40, 0, 0);
                     rock.body.collideWorldBounds = true;
-                    rock.body.velocity.x = 0;
-                    rock.body.velocity.y = 0;
-                    rock.body.drag.setTo(150);
+                    rock.body.mass = 80;
+                    rock.body.damping = 0.95;
+                    rock.body.angularDamping = 0.85;
+                    rock.body.setMaterial(rockMaterial);
+                    rock.tint = 0x1f1b34;
                 }, this);
 
                 map.createFromObjects('Object Layer', 782, 'demon', 0, true, false, this.enemyGroup);
@@ -93,27 +141,23 @@ angular.module('uiApp')
                     enemy.minY = enemy.initialY - 64;
                     enemy.maxY = enemy.initialY + 64;
                     enemy.isChasing = false;
-                    enemy.anchor.set(0.5);
-                    this.game.physics.arcade.enable(enemy, Phaser.Physics.ARCADE, true);
-                    enemy.body.setSize(32, 32, 0, 0);
-                    enemy.body.bounce.set(1);
+                    //enemy.body.debug = true;
+                    enemy.body.setCircle(11);
                     enemy.body.collideWorldBounds = true;
-                    enemy.body.velocity.set(25);
+                    enemy.body.fixedRotation = true;
+                    enemy.body.velocity.x = 25;
+                    enemy.body.velocity.y = 25;
+                    enemy.body.setZeroDamping();
+                    enemy.body.setMaterial(demonMaterial);
+                    enemy.tint = 0x1f1b34;
                 }, this);
 
-                this.game.camera.follow(this.player);
-
+                this.player.body.onBeginContact.add(this.death, this);
                 this.cursors = this.game.input.keyboard.createCursorKeys();
             },
 
             update: function () {
-                this.game.physics.arcade.collide(this.rockGroup, this.blockLayer);
-                this.game.physics.arcade.collide(this.player, this.blockLayer);
-                this.game.physics.arcade.collide(this.enemyGroup, this.blockLayer);
-                this.game.physics.arcade.collide(this.enemyGroup, this.rockGroup, this.noRockMove, undefined, this);
-                this.game.physics.arcade.collide(this.player, this.rockGroup);
-                this.game.physics.arcade.collide(this.player, this.enemyGroup, this.death, undefined, this);
-                this.player.body.velocity.set(0);
+                this.player.body.setZeroVelocity();
                 angular.forEach(this.enemyGroup.children, function (enemy) {
                     if (enemy.isChasing) {
                         //  TODO
@@ -128,44 +172,56 @@ angular.module('uiApp')
                             (compareY >= enemy.maxY && enemy.body.velocity.y > 0)) {
                             enemy.body.velocity.y *= -1;
                         }
+                        if (Math.abs(enemy.body.velocity.x) < 5) {
+                            enemy.body.velocity.x = Math.sign(enemy.body.velocity.x) * -1 * 25;
+                        }
+                        if (Math.abs(enemy.body.velocity.y) < 5) {
+                            enemy.body.velocity.y = Math.sign(enemy.body.velocity.y) * -1 * 25;
+                        }
                     }
                 }, this);
                 if (this.cursors.up.isDown) {
-                    this.player.body.velocity.y = -75;
+                    this.player.body.moveUp(75);
                 }
                 if (this.cursors.down.isDown) {
-                    this.player.body.velocity.y = 75;
+                    this.player.body.moveDown(75);
                 }
                 if (this.cursors.left.isDown) {
-                    this.player.body.velocity.x = -75;
+                    this.player.body.moveLeft(75);
                 }
                 if (this.cursors.right.isDown) {
-                    this.player.body.velocity.x = 75;
+                    this.player.body.moveRight(75);
                 }
             },
 
             render: function () {
-                /*
-                this.game.debug.body(this.player);
-                this.game.debug.cameraInfo(game.camera);
+/*
+                this.game.debug.cameraInfo(game.camera, 0, 0);
+                this.game.debug.spriteInfo(this.player, 400, 0);
                 this.blockLayer.debug = true;
-                angular.forEach(this.enemyGroup.children, function (enemy) {
-                    this.game.debug.body(enemy);
+                var count = 0;
+                angular.forEach(this.enemyGroup.children, function (child, index) {
+                    this.game.debug.spriteInfo(child, index * 350, 100);
+                    ++count;
                 }, this);
-                angular.forEach(this.rockGroup.children, function (enemy) {
-                    this.game.debug.body(enemy);
+                count = 0;
+                angular.forEach(this.rockGroup.children, function (child, index) {
+                    this.game.debug.spriteInfo(child, index * 350, 200);
+                    ++count;
                 }, this);
-                */
+*/
             },
 
-            death: function () {
-                this.player.kill();
-                this.game.state.start('TitleScreen');
-            },
-
-            noRockMove: function(sprite1, sprite2) {
-                sprite2.body.velocity.set(0);
-                sprite1.body.velocity.set(25);
+            death: function (body) {
+                if (
+                    angular.isDefined(body) &&
+                    body !== null &&
+                    angular.isDefined(body.sprite) &&
+                    body.sprite !== null &&
+                    body.sprite.key === 'demon') {
+                    this.player.kill();
+                    this.game.state.start('TestMaze');
+                }
             }
         };
